refactor(networking): move node-only logic to node/ and inject isServer
