+++ tests/vitest.config.ts
@@ -15,9 +15,4 @@ export default defineConfig({
       '#src': r('../src'),
     },
   },
-  // matches the Nitro/server build's values, since all unit tests here exercise server-side code
-  define: {
-    'import.meta.server': 'true',
-    'import.meta.client': 'false',
-  },
 })
